adding V2 version, revise preprocessing and feature engineering, change model to decision tree
