+++ .gitignore
@@ -1,3 +1,5 @@
 Dataset/
 .venv_e
 plds_team_e
+glacial-garden-12652
+data
