--- app/config/feature_config.yaml
@@ -1,1 +1,0 @@
-out_path : "../output/"
