+++ app/config/preprocessing_config.yaml
@@ -1,40 +1,1 @@
-x_col : ['ID', 'LIMIT_BAL', 'SEX', 'EDUCATION', 'MARRIAGE', 'AGE', 'PAY_0', 'PAY_2', 'PAY_3', 'PAY_4', 'PAY_5', 'PAY_6', 'BILL_AMT1', 'BILL_AMT2', 'BILL_AMT3', 'BILL_AMT4', 'BILL_AMT5', 'BILL_AMT6', 'PAY_AMT1', 'PAY_AMT2', 'PAY_AMT3', 'PAY_AMT4', 'PAY_AMT5', 'PAY_AMT6']
-y_col : ['TARGET']
-
 out_path : "../output/"
-
-NUM_COLUMN:
-- LIMIT_BAL
-- AGE
-- PAY_0
-- PAY_2
-- PAY_3
-- PAY_4
-- PAY_5
-- PAY_6
-- BILL_AMT1
-- BILL_AMT2
-- BILL_AMT3
-- BILL_AMT4
-- BILL_AMT5
-- BILL_AMT6
-- PAY_AMT1
-- PAY_AMT2
-- PAY_AMT3
-- PAY_AMT4
-- PAY_AMT5
-- PAY_AMT6
-- Closeness_6
-- Closeness_5
-- Closeness_4
-- Closeness_3
-- Closeness_2
-- Closeness_1
-
-CAT_COLUMN:
-- SEX
-- MARRIAGE
-
-LABEL_COLUMN:
-- EDUCATION
-- AgeBin
