+++ app/src/train.sh
@@ -1,6 +1,6 @@
 python3 read_data.py
-python3 preprocessing.py
-python3 feature_engineering.py
+python3 preprocessing_v2.py
+python3 feature_engineering_v2.py
 python3 modelling.py
 python3 predict.py
 python3 api.py
